Add more installation options for ZSH
- Option to install ZSH through scripts
- Only ask about oh-my-zsh if ZSH is already installed

+++ install.sh
@@ -35,27 +35,31 @@ echo ""
 
 # -- ZSH Setup -----------------------------------------------------------------
 
-if get_boolean_response "Do you want to install ZSH configuration files?"; then
+if exists "zsh"; then
+  if get_boolean_response "Do you want to install ZSH configuration files?"; then
 
-  # -- ZSHRC
-  ln -sf $HOME/.dotfiles/zshrc $HOME/.zshrc
-  echo_item "Linked zshrc" "green"
+    # -- ZSHRC
+    ln -sf $HOME/.dotfiles/zshrc $HOME/.zshrc
+    echo_item "Linked zshrc" "green"
 
-  # -- OH MY ZSH
-  if [ -d $HOME/.oh-my-zsh/ ]; then
-    echo_item "Oh my ZSH is already installed" "green"
-  else
-    if exists "curl"; then
-      curl -L http://install.ohmyz.sh | sh
-    elif exists "wget"; then
-      wget --no-check-certificate http://install.ohmyz.sh -O - | sh
+    # -- OH MY ZSH
+    if [ -d $HOME/.oh-my-zsh/ ]; then
+      echo_item "Oh my ZSH is already installed" "green"
     else
-      echo_item "You need either curl or wget installed to download Oh My ZSH"
+      if exists "curl"; then
+        curl -L http://install.ohmyz.sh | sh
+      elif exists "wget"; then
+        wget --no-check-certificate http://install.ohmyz.sh -O - | sh
+      else
+        echo_item "You need either curl or wget installed to download Oh My ZSH"
+      fi
     fi
-  fi
 
+  else
+    echo_item "Ignoring ZSH configuration" "red"
+  fi
 else
-  echo_item "Ignoring ZSH configuration" "red"
+  echo_item "ZSH is not installed"
 fi
 
 echo ""
